% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospitalizationInSentHomePatients.R
\docType{package}
\name{HospitalizationInSentHomePatients}
\alias{HospitalizationInSentHomePatients}
\title{HospitalizationInSentHomePatients}
\description{
HospitalizationInSentHomePatients
}
